ILSC_MRD_suite_15_modulos_CANON/requirements.txt
numpy>=1.24
PyYAML>=6.0
jsonschema>=4.18
reportlab>=3.6
scipy>=1.10  # optional (audit version capture)
